:recycle: api: separate rescoring in two usecases
